Reorganized solution folders. Renamed services integration test project Create web app integration test project
