+++ DFC.App.Help.PageService.IntegrationTests/ServiceTests/HelpPageServiceTests/HelpPageServiceDeleteTests.cs
@@ -7,7 +7,7 @@ using NUnit.Framework;
 using System;
 using System.Threading.Tasks;
 
-namespace DFC.App.Help.IntegrationTests.ServiceTests.HelpPageServiceTests
+namespace DFC.App.Help.PageService.IntegrationTests.ServiceTests.HelpPageServiceTests
 {
     [TestFixture]
     public class HelpPageServiceDeleteTests : BaseHelpPageServiceTests
