+++ DFC.App.Help.PageService.IntegrationTests/ServiceTests/HelpPageServiceTests/HelpPageServiceGetAllTests.cs
@@ -8,7 +8,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace DFC.App.Help.IntegrationTests.ServiceTests.HelpPageServiceTests
+namespace DFC.App.Help.PageService.IntegrationTests.ServiceTests.HelpPageServiceTests
 {
     [TestFixture]
     public class HelpPageServiceGetAllTests : BaseHelpPageServiceTests
