+++ DFC.App.Help.PageService.IntegrationTests/ServiceTests/HelpPageServiceTests/HelpPageServiceGetByIdTests.cs
@@ -6,17 +6,17 @@ using NUnit.Framework;
 using System;
 using System.Threading.Tasks;
 
-namespace DFC.App.Help.IntegrationTests.ServiceTests.HelpPageServiceTests
+namespace DFC.App.Help.PageService.IntegrationTests.ServiceTests.HelpPageServiceTests
 {
     [TestFixture]
-    public class HelpPageServiceCreateTests : BaseHelpPageServiceTests
+    public class HelpPageServiceGetByIdTests : BaseHelpPageServiceTests
     {
         [Test]
-        [Category("HelpPageService.Create")]
-        public async Task HelpPageServiceCreateReturnsSuccessWhenHelpPageCreated()
+        [Category("HelpPageService.GetById")]
+        public async Task HelpPageServiceGetByIdReturnsSuccessWhenHelpPageExists()
         {
             // arrange
-            const string name = ValidNameValue + "_Create";
+            const string name = ValidNameValue + "_GetById";
             var helpPageModel = new HelpPageModel()
             {
                 CanonicalName = name + "_" + Guid.NewGuid().ToString(),
@@ -24,15 +24,29 @@ namespace DFC.App.Help.IntegrationTests.ServiceTests.HelpPageServiceTests
             };
             var helpPageService = ServiceProvider.GetService<IHelpPageService>();
 
-            // act
             await helpPageService.CreateAsync(helpPageModel).ConfigureAwait(false);
 
+            // act
             var result = await helpPageService.GetByIdAsync(helpPageModel.DocumentId).ConfigureAwait(false);
 
             // assert
-            result.Should().NotBeNull();
             result.DocumentId.Should().Be(helpPageModel.DocumentId);
+            result.Should().NotBeNull();
             result.CanonicalName.Should().Be(helpPageModel.CanonicalName);
+        }
+
+        [Test]
+        [Category("HelpPageService.GetById")]
+        public async Task HelpPageServiceGetByIdReturnsNullWhenHelpPageDoesNotExist()
+        {
+            // arrange
+            var helpPageService = ServiceProvider.GetService<IHelpPageService>();
+
+            // act
+            var result = await helpPageService.GetByIdAsync(Guid.NewGuid()).ConfigureAwait(false);
+
+            // assert
+            result.Should().BeNull();
         }
     }
 }
