+++ DFC.App.Help.PageService.IntegrationTests/ServiceTests/HelpPageServiceTests/HelpPageServicePingTests.cs
@@ -4,7 +4,7 @@ using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 using System.Threading.Tasks;
 
-namespace DFC.App.Help.IntegrationTests.ServiceTests.HelpPageServiceTests
+namespace DFC.App.Help.PageService.IntegrationTests.ServiceTests.HelpPageServiceTests
 {
     [TestFixture]
     public class HelpPageServicePingTests : BaseHelpPageServiceTests
