+++ DFC.App.Help.PageService.IntegrationTests/appsettings-template.json
@@ -1,10 +1,4 @@
 {
-  "Logging": {
-    "LogLevel": {
-      "Default": "Debug"
-    }
-  },
-  "AllowedHosts": "*",
   "Configuration": {
     "CosmosDbConnections": {
       "HelpPages": {
