+++ DFC.App.Help.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.28922.388
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DFC.App.Help", "DFC.App.Help\DFC.App.Help.csproj", "{0A9E48D3-16BF-4E3A-8178-A3E737E270EE}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DFC.App.Help.IntegrationTests", "DFC.App.Help.IntegrationTests\DFC.App.Help.IntegrationTests.csproj", "{5C7A84AC-56E4-4A71-8E3C-12195A2BBE98}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DFC.App.Help.UnitTests", "DFC.App.Help.UnitTests\DFC.App.Help.UnitTests.csproj", "{0BC26FE3-7333-4DE8-BE94-E87BA03644A2}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Data", "Data", "{698CBB4A-0FFD-4028-8EC4-1315DAFA1F9A}"
@@ -39,6 +37,14 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DFC.App.DraftHelp.PageServi
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DFC.App.Help.Repository.SitefinityApi", "DFC.App.Help.Repository.SitefinityApi\DFC.App.Help.Repository.SitefinityApi.csproj", "{803EE493-9784-4900-A823-64597246F0BA}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Services", "Services", "{1FA8991D-1E48-453E-B317-94B0079821A7}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Web", "Web", "{BF2740FE-2D18-4C08-BBA4-7402C8858DEB}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DFC.App.Help.PageService.IntegrationTests", "DFC.App.Help.PageService.IntegrationTests\DFC.App.Help.PageService.IntegrationTests.csproj", "{6BB35EE7-7E2F-44DA-9BFA-4EBA65F21834}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DFC.App.Help.IntegrationTests", "DFC.App.Help.IntegrationTests\DFC.App.Help.IntegrationTests.csproj", "{05E71982-D936-4E36-A2CE-67AF3CEE744F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -49,10 +55,6 @@ Global
 		{0A9E48D3-16BF-4E3A-8178-A3E737E270EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0A9E48D3-16BF-4E3A-8178-A3E737E270EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0A9E48D3-16BF-4E3A-8178-A3E737E270EE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5C7A84AC-56E4-4A71-8E3C-12195A2BBE98}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5C7A84AC-56E4-4A71-8E3C-12195A2BBE98}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5C7A84AC-56E4-4A71-8E3C-12195A2BBE98}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5C7A84AC-56E4-4A71-8E3C-12195A2BBE98}.Release|Any CPU.Build.0 = Release|Any CPU
 		{0BC26FE3-7333-4DE8-BE94-E87BA03644A2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0BC26FE3-7333-4DE8-BE94-E87BA03644A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0BC26FE3-7333-4DE8-BE94-E87BA03644A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -85,12 +87,19 @@ Global
 		{803EE493-9784-4900-A823-64597246F0BA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{803EE493-9784-4900-A823-64597246F0BA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{803EE493-9784-4900-A823-64597246F0BA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6BB35EE7-7E2F-44DA-9BFA-4EBA65F21834}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6BB35EE7-7E2F-44DA-9BFA-4EBA65F21834}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6BB35EE7-7E2F-44DA-9BFA-4EBA65F21834}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6BB35EE7-7E2F-44DA-9BFA-4EBA65F21834}.Release|Any CPU.Build.0 = Release|Any CPU
+		{05E71982-D936-4E36-A2CE-67AF3CEE744F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{05E71982-D936-4E36-A2CE-67AF3CEE744F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{05E71982-D936-4E36-A2CE-67AF3CEE744F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{05E71982-D936-4E36-A2CE-67AF3CEE744F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{5C7A84AC-56E4-4A71-8E3C-12195A2BBE98} = {66F2C5CF-7C1D-42FD-90E3-64DDA31F26E5}
 		{0BC26FE3-7333-4DE8-BE94-E87BA03644A2} = {B8502880-F710-4CDA-88E9-50475C4DE4D8}
 		{2A90123F-762B-4DA7-8917-32239E5411CC} = {C1DA4E73-5D3B-4170-AF0B-ED21AA25CC67}
 		{A513B109-4AFA-4757-ACFD-A81C08CFF675} = {698CBB4A-0FFD-4028-8EC4-1315DAFA1F9A}
@@ -103,6 +112,10 @@ Global
 		{F6A3EB45-844B-45A3-87C8-BF0F034B1717} = {B8502880-F710-4CDA-88E9-50475C4DE4D8}
 		{4338102E-C1AC-4472-8D01-A4AD3BC3F261} = {4653D94B-6CB2-4D42-AF93-81DA13CEDF77}
 		{803EE493-9784-4900-A823-64597246F0BA} = {C1DA4E73-5D3B-4170-AF0B-ED21AA25CC67}
+		{1FA8991D-1E48-453E-B317-94B0079821A7} = {66F2C5CF-7C1D-42FD-90E3-64DDA31F26E5}
+		{BF2740FE-2D18-4C08-BBA4-7402C8858DEB} = {66F2C5CF-7C1D-42FD-90E3-64DDA31F26E5}
+		{6BB35EE7-7E2F-44DA-9BFA-4EBA65F21834} = {1FA8991D-1E48-453E-B317-94B0079821A7}
+		{05E71982-D936-4E36-A2CE-67AF3CEE744F} = {BF2740FE-2D18-4C08-BBA4-7402C8858DEB}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {3581291A-8988-4696-A525-6CDEB71A070C}
